assignment 5
Assignment -5 Q1- Write a multithreaded program that generates the Fibonacci series. This program should work as follows: The user will enter on the command line the number of Fibonacci numbers that the program is to generate. The program will then create a separate thread that will generate the Fibonacci numbers. When the thread finishes execution, the parent thread will output the sequence generated by the child thread. Because the parent thread cannot begin outputting the Fibonacci sequence until the child thread finishes, this will require having the parent thread wait for the child thread to finish. Q2- A C program to show multiple threads with local, global and static variables. Create three threads and print the value of the static, global and local variable.